+++ src/components/Input/SelectInput/SelectInput.component.tsx
@@ -41,7 +41,7 @@ import type { TestIdProps } from "@types";
 /**
  * An input that displays a list of options, and allows the user to select one.
  *
- * @version 0.0.7
+ * @version 0.0.8
  *
  * @param {SelectInputProps & TestIdProps} props - SelectInput component props
  * @returns {JSX.Element}